Typing fixes for mkservice and KIA

+++ src/cmd/ksh93/include/shlex.h
@@ -187,7 +187,7 @@ extern noreturn void 	sh_syntax(Lex_t*, int);
     extern Kia_t		kia;
 
     extern int                  kiaclose(Lex_t *);
-    extern unsigned long        kiaentity(Lex_t*, const char*,int,int,int,int,unsigned long,int,int,const char*);
+    extern unsigned long        kiaentity(Lex_t*, const char*,ssize_t,int,int,int,unsigned long,int,size_t,const char*);
 #endif /* SHOPT_KIA */
 
 #endif /* !NOTSYM */
